+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/service/sdk/wrapper/SendChatMessageResult.kt
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.ui.chat.service.sdk.models
+package com.azure.android.communication.ui.chat.service.sdk.wrapper
 
 internal data class SendChatMessageResult(val id: String)
